+++ packages/relay/src/lib/services/ethService/ethFilterService/index.ts
@@ -1,3 +1,23 @@
+/*-
+ *
+ * Hedera JSON RPC Relay
+ *
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 import { Logger } from 'pino';
 import { ClientCache, MirrorNodeClient } from '../../../clients';
 import constants from '../../../constants';
@@ -43,8 +63,8 @@ export class FilterService implements IFilterService {
 
     const cacheKey = `${constants.CACHE_KEY.FILTERID}_${filterId}`;
     const filter = this.cache.get(cacheKey, this.ethUninstallFilter, requestIdPrefix);
-    
-    if(filter) {
+
+    if (filter) {
       this.cache.delete(cacheKey, this.ethUninstallFilter, requestIdPrefix);
       return true;
     }
